Extract task code into task.ts
